requirements.txt
# ========================================
# Entropy-Regularized QGAN Requirements
# For Single NVIDIA H200 GPU (143GB VRAM)
# ========================================

# Quantum Computing Frameworks
pennylane>=0.35.0
pennylane-lightning-gpu>=0.35.0  # GPU-accelerated quantum simulator
qiskit>=1.0.0
qiskit-aer>=0.13.0

# Deep Learning Frameworks
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0

# Experiment Tracking & Visualization
wandb>=0.16.0
tensorboard>=2.15.0
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# Data & Metrics
scikit-learn>=1.3.0
Pillow>=10.0.0
tqdm>=4.66.0

# Configuration Management
pyyaml>=6.0
omegaconf>=2.3.0

# Utilities
pandas>=2.1.0
jupyter>=1.0.0
ipykernel>=6.26.0
black>=23.0.0  # Code formatting
pytest>=7.4.0  # Testing

# Optional: For IBM Quantum Hardware
# qiskit-ibm-runtime>=0.18.0
# Uncomment if running on real quantum hardware
